LF-5298 Make TimeStrip reusable
